Tests/LibWeb/Text/expected/usvstring-url-reflection.txt

audio.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
embed.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
frame.longDesc final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
frame.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
iframe.longDesc final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
iframe.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
img.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
img.longDesc final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
img.lowsrc final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
input.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
link.href final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
object.codeBase final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
object.data final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
script.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
source.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
track.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
video.poster final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
video.src final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
q.cite final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
blockquote.cite final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
ins.cite final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
del.cite final URL path segment: %EF%BF%BDa%EF%BF%BDb%EF%BF%BD
